feat/booking-initialize

+++ modules/review/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 class Review(models.Model):
     venue = models.ForeignKey(Venue, on_delete=models.CASCADE, related_name='reviews')
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    rating = models.FloatField()
+    rating = models.DecimalField(max_digits=3, decimal_places=1)
     comment = models.TextField(blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
